GEN-30525 - EutPathDB plugin will not find some organisms that are in the database
Add a checkbox to the search UI to select whether query string is a list of gene IDs, or if it is a free text search that can have wildcards and quoted phrases. No longer try to guess that from the content of the query string.
Refactor the EuPathDB service class to remove code duplicated with the child service classes